requirements.txt
python-telegram-bot[job-queue]
libtorrent
telethon
aiohttp
python-dotenv
rarfile
py7zr